PROD/requirements.txt
blinker==1.8.2
branca==0.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
Flask==3.0.2
folium==0.16.0
fonttools==4.51.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
PyMySQL==1.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
tenacity==8.3.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wordcloud==1.9.3
xyzservices==2024.4.0
